move: modal component folder structure edit
