+++ src/components/common/modal/common-modal/CommonAlertModal.stories.tsx
