+++ src/components/common/modal/common-modal/CommonConfirmModal.stories.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import type { Meta, StoryObj } from "@storybook/react-vite";
 import { action } from "storybook/actions";
-import CommonConfirmModal from "./CommonConfirmModal";
+import CommonConfirmModal from "../CommonConfirmModal";
 
 const meta = {
   title: "Components/Modal/CommonConfirmModal",
