+++ src/components/common/modal/full-screen-modal/FullScreenModalContent.tsx
